Geometry.Sculptor.Shapes

Description

Documentation

data Face v

A Face is a list of vertices TODO: Polymorphic container TODO: Make sure all vertices lie in the same plane (using Triangles?)

Constructors

Face [v]

data Edge v

Constructors

Edge v v

vlist :: a -> a -> a -> [a]

A simple list of vertices

triangles :: [a] -> [[a]]

Tessellate a polygon with triangles. Presently, only concave shapes will tessellate properly. TODO: Make it work for any shape (convex and concave) TODO: Opposite rather than adjacent sides (don't include centre)

xy :: Fractional f => (f -> f -> f -> a) -> f -> f -> a

Create a point that lies on the XY plane

xz :: Fractional f => (f -> f -> f -> a) -> f -> f -> a

Create a point that lies on the XZ plane

yz :: Fractional f => (f -> f -> f -> a) -> f -> f -> a

Create a point that lies on the YZ plane

plane :: Fractional f => (f -> f -> a) -> f -> f -> [a]

Generate the vertices for a rectangular plane, centred at (0,0,0) TODO: Rename (eg. rectangle) (?)

Assuming the plane is parallel to the X and Y axes, the vertices are given in a clockwise order, starting from the top left corner (eg. top left, top right, bottom right, bottom left).

planeXY :: Fractional f => (f -> f -> f -> a) -> f -> f -> [a]

Generate the vertices for a rectangular plane, parallel with the X and Y axes and centred at (0,0,0)

planeXZ :: Fractional f => (f -> f -> f -> a) -> f -> f -> [a]

Generate the vertices for a rectangular plane, parallel with the X and Z axes and centred at (0,0,0)

planeYZ :: Fractional f => (f -> f -> f -> a) -> f -> f -> [a]

Generate the vertices for a rectangular plane, parallel with the Y and Z axes and centred at (0,0,0)

polygon :: (Floating f, Integral i) => (f -> f -> a) -> i -> f -> [a]

Generate the vertices for a regular polygon, centred at (0,0,0)

cuboid :: Fractional f => (f -> f -> f -> a) -> f -> f -> f -> [a]

lattice ::

Generate the vertices of an axis-aligned cuboid centred at (0,0,0) TODO: Use combinatorics to generate vertices (?) TODO: Specify order and winding somehow

cube :: Fractional f => (f -> f -> f -> a) -> f -> [a]

Generate the vertices of an axis-aligned cube centred at (0,0,0)

cuboidIndices :: Integral i => [[[i]]]

left, right, bottom, top, back, front V3 top left bottom

Indices for a solid cuboid (with triangulated sides, cf. cuboid) TODO: Use more sensible container types (three levels of lists seems a bit excessive...) TODO: Allow any kind of tessellation (?)

cuboidLineIndices :: Integral i => [(i, i)]